[#16] Added missing test dependency slf4j-simple
Running tests in cloudkeeper-executors/cloudkeeper-forking-executor
triggers error messages that class org.slf4j.impl.StaticLoggerBinder
cannot be loaded. This is due to the fact that some of the packages on
which running of tests depend themselves depend non-transitively on
org.slf4j:slf4j-simple package. This commit fixes this issue.